refactor(forge): own the PR-state record and its reader
Keep harness reads below sidebar policy, with the cache-class read
primitive beside its existing disk publisher.
