+++ crates/rimz/src/forge.rs
@@ -7,6 +7,8 @@ use serde_json::Value;
 
 use crate::store::snapshot::{WorktreePrCi, WorktreePrState};
 
+pub mod pr_state;
+
 const GH_HEAD_FIELDS: &str = "headRefName,headRepository,headRepositoryOwner,isCrossRepository";
 const TEA_LIST_FIELDS: &str = "index,state,head,created";
 const LIST_LIMIT: &str = "500";
