+++ crates/rimz/src/sidebar/refresh/pr/transitions.rs
@@ -4,8 +4,9 @@ use std::collections::BTreeMap;
 
 use serde_json::{Map, Value};
 
-use super::{PrLink, PrStateCache, RepoGroup, TargetStamp};
+use super::RepoGroup;
 use crate::forge::RemoteRepo;
+use crate::forge::pr_state::{PrLink, PrStateCache, TargetStamp};
 use crate::harness::schedule::signal::Signal;
 use crate::store::event::SignalSource;
 use crate::store::snapshot::{WorktreePrCi, WorktreePrState};
